feat: implement modules.js - task 9
